requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
emoji==2.14.1
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
Js2Py_3.13==0.74.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
lxml==6.0.2
m3u8==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.3.0
networkx==3.5
nltk==3.9.1
ntscraper==0.3.18
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
protobuf==6.32.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyjsparser==2.7.1
pyotp==2.9.0
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
safetensors==0.6.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.0
tweepy==4.16.0
twikit==2.3.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
webvtt-py==0.5.1
